Fixes bug caused by trying to acccess delegateTarget from event asynchronously
The `delegateTarget` event property can change after an event has been fired, and it's now being removed after the event finishes being delegated, so it should not be referenced inside async functions.
I've changed the code to store the value of the delegateTarget in a local variable and access that instead inside the async functions.